requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
mysqlclient>=1.3
pillow
django-ckeditor
django-cleanup
python-magic
django-ffmpeg
django-imagefit